requirements.txt
fastapi==0.109.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
celery==5.3.6
redis==5.0.1
python-dotenv==1.0.0
pytest==7.4.3
httpx==0.25.2
pydantic-settings~=2.7.0
pytest-asyncio==0.20.3
pytest-cov==4.1.0
pydantic~=2.10.3